+++ web/js/metadata/Pagemarks.ts
@@ -138,10 +138,15 @@ export class Pagemarks {
 
             for (const pageNum of Numbers.range(start, end)) {
 
-                const rect = createPagemarkRect(pageNum, pageNum === end ? percentage : 100);
+                const rectPercentage =
+                    pageNum === end ? percentage : 100;
+
+                const rect = createPagemarkRect(pageNum, rectPercentage);
 
                 if (rect) {
+
                     const pagemark = Pagemarks.create({created, rect, batch});
+
                     Pagemarks.updatePagemark(docMeta, pageNum, pagemark);
 
                     result.push({pageNum, pagemark});