+++ web/spectron/ui-components/App.tsx
@@ -20,26 +20,21 @@ import {RelatedTags} from '../../js/tags/related/RelatedTags';
 import {CommentIcon} from '../../js/ui/standard_icons/CommentIcon';
 import {FlashcardIcon} from '../../js/ui/standard_icons/FlashcardIcon';
 import {AnnotationFlashcardBox} from '../../js/annotation_sidebar/flashcard_input/AnnotationFlashcardBox';
-import {FlashcardType} from '../../js/metadata/FlashcardType';
 import {FlashcardInputForCloze} from '../../js/annotation_sidebar/flashcard_input/FlashcardInputForCloze';
 import {FlashcardInputForFrontAndBack} from '../../js/annotation_sidebar/flashcard_input/FlashcardInputForFrontAndBack';
-import {TopPDFExample} from './TopPDFExample';
 import {AnnotationCommentBox} from '../../js/annotation_sidebar/AnnotationCommentBox';
 import {DropdownItem, DropdownMenu, DropdownToggle, UncontrolledDropdown} from 'reactstrap';
-import {GDPRNotice} from '../../js/ui/gdpr/GDPRNotice';
 import {ExportButton} from '../../js/ui/export/ExportButton';
 import {EditorsPicksContent} from '../../../apps/repository/js/editors_picks/EditorsPicksContent';
 import {AnkiReviewContent} from './AnkiReviewContent';
 import ReadingProgressTable from '../../../apps/repository/js/stats/ReadingProgressTable';
 import {ContextMenuWrapper, prepareContextMenuHandlers} from '@burtonator/react-context-menu-wrapper';
-import Dropdown from 'reactstrap/lib/Dropdown';
-import {DropMenu} from './DropMenu';
 import {TestMenu} from './TestMenu';
-import Joyride from 'react-joyride';
 import {Feedback} from '../../js/ui/feedback/Feedback';
-import {Survey} from '../../../apps/repository/js/splash/splashes/survey/Survey';
-import {Premium} from '../../../apps/repository/js/splash/splashes/premium/Premium';
-import {ChromeExtensionReview} from '../../../apps/repository/js/splash/splashes/chrome_extension_review/ChromeExtensionReview';
+import Button from 'reactstrap/lib/Button';
+import {Toaster} from '../../js/ui/toaster/Toaster';
+import {ipcRenderer} from 'electron';
+import {AlternativeToReview} from '../../../apps/repository/js/splash/splashes/alternativeto_review/AlternativeToReview';
 
 class App<P> extends React.Component<{}, IAppState> {
 
@@ -135,6 +130,15 @@ class App<P> extends React.Component<{}, IAppState> {
                 content: 'This another awesome feature!',
             },
         ];
+        // Toaster.success('A new update for Polar was downloaded.  Please restart.', 'Update downloaded', {
+        //          requiresAcknowledgment: true,
+        //          preventDuplicates: true
+        //      });
+        //
+        // Toaster.info('X A new update for Polar was downloaded.  Please restart.', 'Update downloaded', {
+        //     requiresAcknowledgment: true,
+        //     preventDuplicates: true
+        // });
 
         return (
 
@@ -142,7 +146,64 @@ class App<P> extends React.Component<{}, IAppState> {
 
                 {/*<ChromeExtensionReview settingKey={'asdf'}/>*/}
 
-                <Premium settingKey={'premium-key'}/>
+                {/*<div style={{*/}
+                        {/*width: '500px',*/}
+                        {/*position: 'fixed',*/}
+                        {/*right: 10,*/}
+                        {/*bottom: 10,*/}
+                        {/*zIndex: 9999,*/}
+                    {/*}}*/}
+                    {/*className="border rounded shadow p-3 m-2 text-white bg-dark">*/}
+
+                    {/*<div style={{*/}
+                            {/*display: 'flex',*/}
+                            {/*verticalAlign: 'middle'*/}
+                        {/*}}*/}
+                        {/*className="mb-3">*/}
+
+                        {/*<div className="mr-3 text-success mt-auto mb-auto">*/}
+
+                            {/*<i style={{fontSize: '50px'}} className="fas fa-check"></i>*/}
+
+                        {/*</div>*/}
+
+                        {/*<div className="mt-1 mb-1">*/}
+
+                            {/*<div className="mb-1" style={{fontSize: '18px'}}>*/}
+                                {/*<b>Update available.</b> Please restart.*/}
+                            {/*</div>*/}
+
+                            {/*<div className="mt-1 mb-1 h6">*/}
+                                {/*An update was downloaded and ready to be*/}
+                                {/*installed. Please restart to install the latest*/}
+                                {/*version.*/}
+                            {/*</div>*/}
+
+                        {/*</div>*/}
+
+                    {/*</div>*/}
+
+                    {/*<div>*/}
+
+                        {/*<div className="text-center text-white">*/}
+                            {/*<Button onClick={() => ipcRenderer.send('app-update:quit-and-install')}*/}
+                                    {/*style={{*/}
+                                        {/*fontWeight: 'bold'*/}
+                                    {/*}}*/}
+                                    {/*size="lg"*/}
+                                    {/*color="success">*/}
+                                {/*Restart*/}
+                            {/*</Button>*/}
+                        {/*</div>*/}
+
+                    {/*</div>*/}
+
+
+                {/*</div>*/}
+
+                <AlternativeToReview settingKey={'asdf'}/>
+
+                {/*<Premium settingKey='premium-key'/>*/}
 
                 <h1 className="component">Feedback without description</h1>